requirements.txt
fastapi
uvicorn
httpx
requests
pydantic
pytest
black
flake8
python-dotenv